Changed function arguments from options to exactly one always
CLR entry points have to be either no arguments or a vector of strings. We don't have vectors, so I had been using an option type to represent zero (entry points, mostly) or one (user functions, mostly) arguments. Instead I now have an explicit "nil" () for the entry point case. You can't really do anything with nil, and it can't be converted to anything, but it serves well enough as the argument to the entry point representing no arguments, and it doesn't change the language syntax at all. Doing this allows a lot of simplification of the compiler code.